refactor: move symbol expansion into its own writer_symbols module
Most of this PR's additions to agent.py were symbol-expansion machinery
(169 of 257 inserted lines in one hunk), pushing an already long module
past 1000 lines. Extract it to cofacts_ai/writer_symbols.py, following
media_filedata.py: a module docstring carrying the "why", the private
constants with their reasoning, and the callback itself.

agent.py drops 1098 -> 936 lines and keeps only the wiring
(before_tool_callback=expand_writer_symbols). Tests split the same way:
the 18 symbol cases move to tests/test_writer_symbols.py, leaving
test_writer_callbacks.py to after_tool and handle_writer_tool_error.

Pure code motion -- no behaviour change, same 55 tests pass. Comments
that said "below"/"above" across the old single-file layout are
rewritten to name the module they now point at, and the ADR plus
docs/index.md are updated to the new paths.
